Fixed computation of ip mask

+++ src/httpserver/http_utils.hpp
@@ -302,7 +302,7 @@ struct ip_representation
     int weight() const
     {
         //variable-precision SWAR algorithm
-        unsigned int x = mask;
+        unsigned short x = mask;
         x = x - ((x >> 1) & 0x55555555);
         x = (x & 0x33333333) + ((x >> 2) & 0x33333333);
         return (((x + (x >> 4)) & 0x0F0F0F0F) * 0x01010101) >> 24;
